+++ codecov.yml
@@ -1,0 +1,4 @@
+ignore:
+  - "pipestat/argparser.py"
+  - "pipestat/cli.py"
+  - "pipestat/__main__.py"
