+++ pipestat/__main__.py
@@ -1,4 +1,4 @@
-from .pipestat import main
+from .cli import main
 import sys
 
 if __name__ == "__main__":
